+++ reserve-service/src/main/java/com/bit/docker/reserveservice/infra/repository/ReservationRepository.java
@@ -3,5 +3,5 @@ package com.bit.docker.reserveservice.infra.repository;
 import com.bit.docker.reserveservice.domain.entity.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
 }
